Unify max samples stored tests
